project/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Django Rest Framework boilerplate"
authors = ["AleksaC <aleksacukovic1@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1, <3.9"
django = "^3.0.6"
psycopg2 = "^2.8.5"
djangorestframework = "^3.11.0"
django-redis = "^4.11.0"
gunicorn = "^20.0.4"
django-health-check = "^3.12.1"
djangorestframework-simplejwt = "^4.4.0"
django-celery-beat = "^2.0.0"
django-celery-results = "^1.2.1"
django-cors-headers = "^3.3.0"
django-storages = "^1.9.1"
boto3 = "^1.13.13"
django-anymail = {extras = ["mailgun"], version = "^7.1.0"}

[tool.poetry.dev-dependencies]
coverage = "^5.1"
django-debug-toolbar = "^2.2"
toml = "^0.10.1"
drf-yasg = "^1.17.1"
pre-commit = "^2.4.0"
coveralls = "^2.0.0"

[tool.black]
line-length = 91
exclude = "migrations"

[tool.coverage.run]
branch = true
source = ["."]

omit = [
    "*/migrations/*",
    "*/settings/*",
    "*/wsgi.py",
    "*/asgi.py",
    "manage.py",
    "tests/*",
    "*/venv/*",
]

[tool.coverage.report]
fail_under = 100
show_missing = true
skip_covered = true

exclude_lines = [
    "pragma: no cover",
    "except ImportError",
    "raise ImportError",
    "raise AssertionError",
    "raise NotImplementedError",
    "if settings.DEBUG"
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
